[level 1] Title: 폰켓몬, Time: 3.22 ms, Memory: 89.8 MB -BaekjoonHub

+++ 프로그래머스/1/1845. 폰켓몬/폰켓몬.java
@@ -1,10 +1,10 @@
 import java.util.HashSet;
 
 class Solution {
-    public static long solution(int[] nums) {
-        long answer = 0;
-        HashSet<Long> hashSet = new HashSet<>();
-        for(long kind : nums) {
+    public static int solution(int[] nums) {
+        int answer = 0;
+        HashSet<Integer> hashSet = new HashSet<>();
+        for(int kind : nums) {
             hashSet.add(kind);
         }
         if(hashSet.size() >= nums.length / 2) {
